Split up Result into smaller components
