+++ web/src/components/reply/DefReply.svelte
@@ -1,0 +1,29 @@
+<script lang="typescript">
+  import * as reply from "../../util/reply";
+  import NumberParts from "../NumberParts.svelte";
+  import Expr from "../Expr.svelte";
+
+  export let value: reply.DefReply;
+</script>
+
+<h3>{value.canonName}</h3>
+
+{#if value.defExpr}
+  <p>
+    Definition:
+    <Expr value={value.defExpr.ast} />
+  </p>
+{:else if value.type == 'def'}
+  <p>Definition: {value.def}</p>
+{/if}
+
+{#if value.defExpr && value.value}
+  <p>
+    Value:
+    <NumberParts number={value.value} />
+  </p>
+{:else if value.defExpr}
+  <p>Value: {value.defExpr}</p>
+{/if}
+
+<p>{value.doc}</p>
