+++ web/src/components/reply/GenericError.svelte
@@ -1,0 +1,7 @@
+<script lang="typescript">
+  import * as reply from "../../util/reply";
+
+  export let value: reply.GenericError;
+</script>
+
+<p>{value.message}</p>
