+++ web/src/components/reply/NumberReply.svelte
@@ -1,0 +1,10 @@
+<script lang="typescript">
+  import * as reply from "../../util/reply";
+  import NumberParts from "../NumberParts.svelte";
+
+  export let value: reply.NumberReply;
+</script>
+
+<p>
+  <NumberParts number={value} />
+</p>
